requirements.txt
praw
Markdown
python-dotenv